+++ CHANGELOG.md
@@ -62,6 +62,8 @@ ssh_authorized_keys
 
 ### Removed
 
+- remove variable *host_ORACLE_ARCH*, *host_ORACLE_DATA* and *host_ORACLE_ROOT*
+
 ## [2.4.0] - 2023-03-29
 
 ### Added